[demo] mem test in Stage2
This demo showcases the use of TraverseMemoryResourceHob
to conduct a basic memory test in stage2.

The test does not examine the memory region: {MemPoolStart, Top of Low MEM}.
To inspect the reserved region, it must save and rebase all data/code in
that area, and then restore the saved data/code back to the reserved
region after the test.

To conduct a memory test in the 4GB+ space, SBL needs to be compiled
in 64-bit mode. e.g., python BuildLoader.py build tgl -a x64

Signed-off-by: Stanley Chang <stanley.chang@intel.com>