[demo] I2C0 Bus Speed Control (Not Recommended for Use)
This patch is only for demo purposes and is not recommended
for actual use. It uses a legacy and non-standard method (_DSD) to
control the speed of the I2C bus.

ACPI has standardized I2C resources (I2cSerialBusV2 and I2cSerialBus),
which can be declared by an I2C device in ACPI to request for a
different I2C speed.

Prior to this standardization, older platforms (e.g., APL) used
device-specific data (_DSD) in the controller to manage I2C speed.

It is recommended to rewrite the legacy _DSD into the standardized I2C
resources for each individual I2C device in ACPI.

Check the following resources to learn how to use a standard approach:

1. ACPI spec:
  https://uefi.org/sites/default/files/resources/ACPI_5_1release.pdf.
  ps: I2cSerialBusV2 is defined in newer acpi spec (e.g., ACPI 6.0).

2. ACPI enumeration in Linux kernel:
  https://docs.kernel.org/firmware-guide/acpi/enumeration.html.

Signed-off-by: Stanley Chang <stanley.chang@intel.com>