[demo] change the IRQ of a PCIe RP
To change the IRQ of a PCIe RP, for IOAPIC, one must modify:
  1. _PRT (AR00 for IOAPIC in PciTree.asl)
  2. FspsConfig->DevIntConfigPtr

  3. changes must apply to the Bus/Device/Func (BDF) shown in
     EDS corresponding to the DID listed by "lspci". It is possible
     that the BDF shown by "lspci" does not match the BDF in EDS,
     because of "root port function swapping"

OS should disable interrupt remapping and MSI. It can be done
in Linux by adding the kernel command line:
  "intremap=off intel_iommu=off pci=nomsi"

Note, it is not required to set PcieRpHotPlug[Index] = 0x1. The change
in this patch is for demo only.

Discussion: https://groups.io/g/slimbootloader/message/2969

Signed-off-by: Stanley Chang <stanley.chang@intel.com>