[test] Benchmark MTRR Write performance by overriding the "mm" command
This test commit shows how to benchmark write performance by reusing
the "mm" command within the OS Loader.  The "mm" command is overridden
to support the following syntax:

  mm [start_address] [val] -w [count] -c [size]

    where,
      "start_address" is the starting memory address for testing,
      "val" is an initial value written to the memory for test,
      "count" is a hexadecimal number larger than "8",
             representing the number of iterations (count-8) the test will execute,
      "size" is the memory size (in hex) to test in each iteration.

    the output format is:

      "done: [time] ms (Size=[size], Count=[count])"

To benchmark the MTRR, it is advisable to first issue the "mtrr" command
to list the MTRR regions. Based on the region information, you can then
perform the MTRR benchmark on different types of MTRR regions. Please be
aware that the test may potentially corrupt the BIOS region in use, which
could result in a crash.

Benchmark Examples:

  Shell> mtrr

  [0x000000000-0x07FFFFFFF] Write-Back
  [0x07C000000-0x07FFFFFFF] Uncacheable
  [0x090000000-0x09FFFFFFF] Write Combining

  # benchmark the write-back region
  Shell> mm 0x30000000 3 -w 100 -c 100000
  done: 28 ms (Size=1048576, Count=248)

  # benchmark the unchcheable region
  Shell> mm 0x7B000000 3 -w 100 -c 100000
  done: 437 ms (Size=1048576, Count=248)

  # benchmark the write-combining region
  Shell> mm 0x90000000 3 -w 100 -c 100000
  done: 36 ms (Size=1048576, Count=248)

Signed-off-by: Stanley Chang <stanley.chang@intel.com>