[demo] TOLUD for 2.5GB, MMIO for 1.5GB, with iGfx enabled
The min requirement for MMIO space under 4GB is:
   req 1: PCIe Space (bus 00-ff): 0xC000_0000 - 0xCFFF_FFFF for TGL
   req 2: MMIO space w/o iGfx Aperture: (about) 0x2D0_0000
   req 3: iGfx MMIO for 256MB Aperture: 0x1000_0000
   req 4: Platform reservation: 0xF000_0000 to 0xFFFF_FFFF
Total space: less than 1GB. Note that req 2 and req 3 need to meet memory alignment, and
as a result, req 2 may occupy a 0x1000_0000 MMIO space.

The commit uses 0xA000_0000 (which 4GB - 1.5GB) for PCI_MEM32_BASE.
Thus, the 0xA000_0000 to 0xBFFF_FFFF can meet the requirement for req 2 and req 3.
When boot with Linux, the /proc/iomap shows,

  a0000000-cfffffff : PCI Bus 0000:00
    a0000000-a0ffffff : input-c-20054a
    (skip)
    b0000000-bfffffff : 0000:00:02.0

Signed-off-by: Stanley Chang <stanley.chang@intel.com>